Integration test for the defect and its subsequently discovered defect. Im using near-identical settings to existing test so i have created initializer test base to not repeat complex chunk of code and im just passing in the DefaultProjectConfig into the base class.
